Refactor highlight_text and share across interfaces
