+++ app.py
@@ -94,98 +94,7 @@ def load_law_articles(dir_path: str = "output") -> dict[str, dict[str, str]]:
 LAW_ARTICLES = load_law_articles()
 
 
-def highlight_text(
-    text: str,
-    entities: list[dict],
-    article_map: dict[str, str] | None = None,
-    ref_targets: dict[str, list[str]] | None = None,
-    tooltips: dict[str, str] | None = None,
-    article_texts: dict[str, str] | None = None,
-) -> str:
-    popup = (
-        '<div id="article-popup" style="display:none;position:fixed;top:10%;'
-        "left:10%;width:80%;max-height:80%;overflow:auto;background-color:white;"
-        'border:1px solid #888;padding:10px;z-index:1000;">'
-        "<button onclick=\"document.getElementById('article-popup').style.display='none';\" style=\"float:left;\">X</button>"
-        '<div id="article-popup-content"></div></div>'
-    )
-
-    parts: list[str] = []
-    last = 0
-    for ent in sorted(entities, key=lambda e: int(e.get("start_char", 0))):
-        try:
-            start = int(ent.get("start_char", 0))
-            end = int(ent.get("end_char", 0))
-        except Exception:
-            continue
-        if (
-            start < last
-            or start < 0
-            or end <= start
-            or start >= len(text)
-            or end > len(text)
-        ):
-            continue
-        parts.append(html.escape(text[last:start]))
-        span_text = html.escape(text[start:end])
-        tooltip = ""
-        rel_attr = ""
-        if tooltips is not None:
-            tip = tooltips.get(str(ent.get("id")))
-            if tip:
-                esc_tip = html.escape(tip)
-                tooltip = f' title="{esc_tip}"'
-                rel_attr = f' data-rel="{esc_tip}"'
-
-        target: str | None = None
-        if ref_targets is not None:
-            targets = ref_targets.get(str(ent.get("id")))
-            if targets:
-                target = str(targets[0])
-
-        if (
-            target is None
-            and ent.get("type") == "ARTICLE"
-            and article_map is not None
-        ):
-            num = canonical_num(ent.get("normalized") or ent.get("text"))
-            if num is not None and num in article_map:
-                target = article_map[num]
-
-        inner = f'<mark id="ent-{ent["id"]}" class="entity-mark"{tooltip}>{span_text}</mark>'
-        art_data = ""
-        if article_texts is not None:
-            art = None
-            if ent.get("type") == "ARTICLE":
-                num = canonical_num(ent.get("normalized") or ent.get("text"))
-                if num is not None:
-                    art = article_texts.get(num)
-            if art is None:
-                art = article_texts.get(f"ID_{ent.get('id')}")
-            if art:
-                art = art.replace("\n", "<br/>")
-                art_data = f' data-article="{html.escape(art)}"'
-
-        if target or rel_attr or art_data:
-            click_js = (
-                "var a=this.getAttribute('data-article');"
-                "var r=this.getAttribute('data-rel');"
-                "var c=a||r;"
-                " if(c){var p=document.getElementById('article-popup');"
-                " var s=document.getElementById('article-popup-content');"
-                " if(s){s.innerHTML=c;} if(p){p.style.display='block';}}"
-            )
-            parts.append(
-                f'<span id="{ent["id"]}" class="ner-span"><a href="javascript:void(0)"{art_data}{rel_attr} onclick="{click_js}">{inner}</a></span>'
-            )
-        else:
-            parts.append(
-                f'<span id="{ent["id"]}" class="ner-span">{inner}</span>'
-            )
-        last = end
-    parts.append(html.escape(text[last:]))
-    html_str = "".join(parts)
-    return popup + f'<div dir="rtl">{html_str}</div>'
+from highlight import highlight_text
 
 
 def build_graph(entities: list[dict], relations: list[dict]) -> str | None:
